+++ index.js
@@ -2,6 +2,7 @@ import AlertController from './app/components/daisy/feedback/alert_controller';
 import CallyInputController from './app/components/daisy/data_input/cally_input_controller';
 import CountdownController from './app/components/daisy/data_display/countdown_controller';
 import ModalController from './app/components/daisy/actions/modal_controller';
+import StickyController from './app/components/loco/sticky_controller';
 import TabsController from './app/components/daisy/navigation/tabs_controller';
 import ThemeController from './app/components/daisy/actions/theme_controller';
 
@@ -10,6 +11,7 @@ export {
   CallyInputController,
   CountdownController,
   ModalController,
+  StickyController,
   TabsController,
   ThemeController,
 };